Feature/circleci and docker (#16)
* Added Dockerfile for creating a Docker container containing a compiled Slicer, with all build files

* ENH: Added Dockerfile for creating container with precompiled Slicer + ALive plugin

* ENH: Use ALive fork of Slicer in Dockerfile

* ENH: Added initial circleci config file. Not finished

* ENH: Updated circleci config file

* ENH: Clone with https instead of git

* Delete Dockerfile

I think this file is not necessary. The idea would be to use the [slicer/slicer-build](https://hub.docker.com/r/slicer/slicer-build) docker image as a base and move the rest of the commands to the circleci configuration

* Delete Dockerfile

I think this is not necessary. We can use the [slicer/slicer-build](https://hub.docker.com/r/slicer/slicer-build) docker image

* ENH: Updated CircleCI config file to use aliveresearch/slicer-build Docker image

* COMP: Fixed bug in CircleCI config

* ENH: Run tests in the slicer-build-ubuntu2004 image

Co-authored-by: Rafael Palomar <RafaelPalomar@users.noreply.github.com>