ENH: Adaptive algorithm for reducing volumetry calculation error
Bezier surface is controlled by 16 points and the diagonal points pairs ([0,5,10,15] and [3,6,9,12]) would create the longest Bezier curve within the Bezier surface. The idea is approximate the two diagonal Bezier lines' length of the bezier surface, then divided by minimal voxel space to get a resolution, so that when the Bezier surface polygon is cutting through voxels we have at least one vertex in each voxel. Then when we label the voxels we can get a safe barrier for region growing.